+++ src/main/java/org/example/repository/PostRepository.java
@@ -4,11 +4,11 @@ import org.example.domain.Post;
 
 import java.util.List;
 
-public interface PostRepository extends CrudRepository<Post, Integer> {
+public interface PostRepository extends CrudRepository<Post, Long> {
 
-    List<Post> findAllByLabelId(Integer labelId);
+    List<Post> findAllByLabelId(Long labelId);
 
-    List<Post> findAllByWriterId(Integer writerId);
+    List<Post> findAllByWriterId(Long writerId);
 
 }
