+++ src/main/java/org/example/repository/WriterRepository.java
@@ -2,5 +2,5 @@ package org.example.repository;
 
 import org.example.domain.Writer;
 
-public interface WriterRepository extends CrudRepository<Writer, Integer> {
+public interface WriterRepository extends CrudRepository<Writer, Long> {
 }
